+++ src/adapters/chartjs.js
@@ -151,7 +151,8 @@ let setFormatOptions = function(chart, options, chartType) {
     thousands: chart.options.thousands,
     decimal: chart.options.decimal,
     precision: chart.options.precision,
-    round: chart.options.round
+    round: chart.options.round,
+    zeros: chart.options.zeros
   };
 
   if (chart.options.bytes) {